Add WidthHeightHitbox class
