+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/ObjectHitboxes/Rectangle.cs
@@ -8,34 +8,14 @@ using System.Threading.Tasks;
 namespace GDEdit.Utilities.Objects.GeometryDash.ObjectHitboxes
 {
     /// <summary>Represents a rectangular hitbox.</summary>
-    public class Rectangle : HitboxType, IHasWidth, IHasHeight
+    public class Rectangle : WidthHeightHitbox, IHasWidth, IHasHeight
     {
-        /// <summary>The width of the rectangular hitbox.</summary>
-        public double Width { get; set; }
-        /// <summary>The height of the rectangular hitbox.</summary>
-        public double Height { get; set; }
-
         /// <summary>Initializes a new instance of the <seealso cref="Rectangle"/> class.</summary>
         /// <param name="both">The length of both dimensions of the rectangular hitbox.</param>
-        public Rectangle(double both) : this(both, both) { }
+        public Rectangle(double both) : base(both) { }
         /// <summary>Initializes a new instance of the <seealso cref="Rectangle"/> class.</summary>
         /// <param name="width">The width of the rectangular hitbox.</param>
         /// <param name="height">The height of the rectangular hitbox.</param>
-        public Rectangle(double width, double height)
-            : base()
-        {
-            Width = width;
-            Height = height;
-        }
-
-        /// <summary>Determines whether a point is within the hitbox.</summary>
-        /// <param name="point">The point's location.</param>
-        /// <param name="hitboxCenter">The hitbox's center.</param>
-        public override bool IsPointWithinHitbox(Point point, Point hitboxCenter)
-        {
-            Point start = new Point(hitboxCenter.X - Width / 2, hitboxCenter.Y - Height / 2);
-            Point end = new Point(hitboxCenter.X + Width / 2, hitboxCenter.Y + Height / 2);
-            return start <= point && point <= end;
-        }
+        public Rectangle(double width, double height) : base(width, height) { }
     }
 }
